Refactor Authorization Endpoint
